+++ Libplanet.Tests/Blocks/BlockMetadataTest.cs
@@ -123,7 +123,7 @@ namespace Libplanet.Tests.Blocks
                 .Add("protocol_version", 4)
                 .Add(
                     "transaction_fingerprint",
-                    ParseHex("104b4243131f490cd1f21f22a142609a8b464c0528ea5fe4d168b9e19f3c4830"));
+                    ParseHex("df39a62cc055e8132163f303e1f21a751d42ba0ace27f1a851ad38fea35a00c4"));
             AssertBencodexEqual(expectedGenesis, GenesisMetadata.MakeCandidateData(default));
             AssertBencodexEqual(
                 expectedGenesis.SetItem("nonce", new byte[] { 0x00, 0x01, 0x02 }),
@@ -239,7 +239,7 @@ namespace Libplanet.Tests.Blocks
 
             HashDigest<SHA256> hash = GenesisMetadata.DerivePreEvaluationHash(default);
             AssertBytesEqual(
-                FromHex("6381a3ac169f38e56a3b7b3a8ebff8927c7667850c36e46433e3499c8dd356d7"),
+                FromHex("bf0906321cc3ee00bc541a7500428ea1c6f6a2e3aa3ba1995790911c588c615c"),
                 hash.ByteArray);
 
             hash = Block1Metadata.DerivePreEvaluationHash(